15686 - Remove Restx (#2494)
* refactored endpoints removing restx

* config clean up after removing restx dependency

* - move v1 endpoints into v1 folder

* fix linting

* - fix sonarcloud identified security issues

* fix sonarcloud identified security hotspots

* update requires prod.txt

* PR Review updates, clean up old code smells
