+++ auth-api/requirements/prod.txt
@@ -6,7 +6,6 @@ Flask-Script
 Flask-Moment
 Flask-SQLAlchemy
 SQLAlchemy-Continuum
-flask-restx
 flask-marshmallow==0.11.0
 flask-jwt-oidc
 python-dotenv
@@ -23,6 +22,7 @@ minio
 pyhumps
 sqlalchemy-utils
 Flask-Caching
+Flask-Cors
 asyncio-nats-client
 asyncio-nats-streaming
 sqlalchemy<1.4
