+++ auth-api/src/auth_api/resources/v1/account.py
@@ -11,33 +11,26 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Endpoints to reset test data from database."""
+"""API endpoints for managing an Org resource."""
 
-from flask import Blueprint
+from flask import Blueprint, request
 from flask_cors import cross_origin
-
 from auth_api import status as http_status
 from auth_api.auth import jwt as _jwt
-from auth_api.exceptions import BusinessException
-from auth_api.services import ResetTestData as ResetService
+from auth_api.services.authorization import Authorization as AuthorizationService
 from auth_api.tracer import Tracer
 from auth_api.utils.endpoints_enums import EndpointEnum
-from auth_api.utils.roles import Role
 
-
-bp = Blueprint('RESET', __name__, url_prefix=f'{EndpointEnum.TEST_API.value}/reset')
+bp = Blueprint('ACCOUNTS', __name__, url_prefix=f'{EndpointEnum.API_V1.value}/accounts')
 TRACER = Tracer.get_instance()
 
 
-@bp.route('', methods=['POST'])
+@bp.route('/<int:account_id>/products/<string:product_code>/authorizations', methods=['GET', 'OPTIONS'])
+@_jwt.requires_auth
 @TRACER.trace()
 @cross_origin(origin='*')
-@_jwt.has_one_of_roles([Role.TESTER.value])
-def post_reset():
-    """Cleanup test data by the provided token."""
-    try:
-        ResetService.reset()
-        response, status = '', http_status.HTTP_204_NO_CONTENT
-    except BusinessException as exception:
-        response, status = {'code': exception.code, 'message': exception.message}, exception.status_code
-    return response, status
+def get_account_product_authorizations(account_id, product_code):
+    """Return authorizations for a product in an account."""
+    expanded: bool = request.args.get('expanded', False)
+    authorizations = AuthorizationService.get_account_authorizations_for_product(account_id, product_code, expanded)
+    return authorizations, http_status.HTTP_200_OK
