+++ auth-api/src/auth_api/resources/v1/codes.py
@@ -11,33 +11,29 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Endpoints to reset test data from database."""
-
-from flask import Blueprint
+"""API endpoints for managing an Invitation resource."""
+from flask import Blueprint, jsonify
 from flask_cors import cross_origin
 
 from auth_api import status as http_status
-from auth_api.auth import jwt as _jwt
 from auth_api.exceptions import BusinessException
-from auth_api.services import ResetTestData as ResetService
-from auth_api.tracer import Tracer
+from auth_api.services import Codes as CodeService
 from auth_api.utils.endpoints_enums import EndpointEnum
-from auth_api.utils.roles import Role
+
+bp = Blueprint('CODES', __name__, url_prefix=f'{EndpointEnum.API_V1.value}/codes')
 
 
-bp = Blueprint('RESET', __name__, url_prefix=f'{EndpointEnum.TEST_API.value}/reset')
-TRACER = Tracer.get_instance()
-
-
-@bp.route('', methods=['POST'])
-@TRACER.trace()
+@bp.route('/<string:code_type>', methods=['GET', 'OPTIONS'])
 @cross_origin(origin='*')
-@_jwt.has_one_of_roles([Role.TESTER.value])
-def post_reset():
-    """Cleanup test data by the provided token."""
+def get_codes(code_type):
+    """Return the codes by giving name."""
     try:
-        ResetService.reset()
-        response, status = '', http_status.HTTP_204_NO_CONTENT
+        codes = CodeService.fetch_codes(code_type=code_type)
+        if codes is not None:
+            response, status = jsonify(codes), http_status.HTTP_200_OK
+        else:
+            response, status = jsonify({'message': f'The code type ({code_type}) could not be found.'}), \
+                http_status.HTTP_404_NOT_FOUND
     except BusinessException as exception:
         response, status = {'code': exception.code, 'message': exception.message}, exception.status_code
     return response, status
