+++ auth-api/src/auth_api/resources/v1/documents_affidavit.py
@@ -11,33 +11,38 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Endpoints to reset test data from database."""
+"""API endpoints for managing a document resource."""
 
-from flask import Blueprint
+
+from flask import Blueprint, send_from_directory
 from flask_cors import cross_origin
 
 from auth_api import status as http_status
-from auth_api.auth import jwt as _jwt
 from auth_api.exceptions import BusinessException
-from auth_api.services import ResetTestData as ResetService
+from auth_api.services import Documents as DocumentService
 from auth_api.tracer import Tracer
 from auth_api.utils.endpoints_enums import EndpointEnum
-from auth_api.utils.roles import Role
+from auth_api.utils.enums import ContentType, DocumentType
 
 
-bp = Blueprint('RESET', __name__, url_prefix=f'{EndpointEnum.TEST_API.value}/reset')
+bp = Blueprint('DOCUMENTS_AFFIDAVIT', __name__, url_prefix=f'{EndpointEnum.API_V1.value}/documents/affidavit')
 TRACER = Tracer.get_instance()
 
 
-@bp.route('', methods=['POST'])
+@bp.route('', methods=['GET', 'OPTIONS'])
 @TRACER.trace()
 @cross_origin(origin='*')
-@_jwt.has_one_of_roles([Role.TESTER.value])
-def post_reset():
-    """Cleanup test data by the provided token."""
+def get_affidavit_document():
+    """Return the Affidavit."""
     try:
-        ResetService.reset()
-        response, status = '', http_status.HTTP_204_NO_CONTENT
+        doc = DocumentService.fetch_latest_document(DocumentType.AFFIDAVIT.value)
+        if doc is None:
+            response, status = {'message': 'The requested document could not be found.'},\
+                http_status.HTTP_404_NOT_FOUND
+        elif doc.as_dict().get('content_type', None) == ContentType.PDF.value:  # pdf has to be served as attachment
+            return send_from_directory('static', filename=doc.as_dict()['content'], as_attachment=True)
+        else:
+            response, status = doc.as_dict(), http_status.HTTP_200_OK
     except BusinessException as exception:
         response, status = {'code': exception.code, 'message': exception.message}, exception.status_code
     return response, status
