+++ auth-api/src/auth_api/resources/v1/notifications.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Endpoints to reset test data from database."""
+"""API endpoints for managing an Notification resource."""
 
 from flask import Blueprint
 from flask_cors import cross_origin
@@ -19,25 +19,26 @@ from flask_cors import cross_origin
 from auth_api import status as http_status
 from auth_api.auth import jwt as _jwt
 from auth_api.exceptions import BusinessException
-from auth_api.services import ResetTestData as ResetService
+from auth_api.services import Membership as MembershipService
 from auth_api.tracer import Tracer
 from auth_api.utils.endpoints_enums import EndpointEnum
 from auth_api.utils.roles import Role
 
 
-bp = Blueprint('RESET', __name__, url_prefix=f'{EndpointEnum.TEST_API.value}/reset')
+bp = Blueprint('NOTIFICATIONS', __name__,
+               url_prefix=f'{EndpointEnum.API_V1.value}/users/<string:user_id>/org/<int:org_id>/notifications')
 TRACER = Tracer.get_instance()
 
 
-@bp.route('', methods=['POST'])
+@bp.route('', methods=['GET', 'OPTIONS'])
 @TRACER.trace()
 @cross_origin(origin='*')
-@_jwt.has_one_of_roles([Role.TESTER.value])
-def post_reset():
-    """Cleanup test data by the provided token."""
+@_jwt.has_one_of_roles([Role.SYSTEM.value, Role.STAFF.value, Role.PUBLIC_USER.value])
+def get_notifications(user_id, org_id):  # pylint:disable=unused-argument
+    """Find the count of notification remaining.If any details invalid, it returns zero."""
     try:
-        ResetService.reset()
-        response, status = '', http_status.HTTP_204_NO_CONTENT
+        pending_count = MembershipService.get_pending_member_count_for_org(org_id)
+        response, status = {'count': pending_count}, http_status.HTTP_200_OK
     except BusinessException as exception:
         response, status = {'code': exception.code, 'message': exception.message}, exception.status_code
     return response, status
