+++ auth-api/src/auth_api/resources/v1/permissions.py
@@ -11,33 +11,40 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Endpoints to reset test data from database."""
+"""API endpoints for managing a Product resource."""
 
-from flask import Blueprint
+import json
+
+from flask import Blueprint, request
 from flask_cors import cross_origin
 
 from auth_api import status as http_status
 from auth_api.auth import jwt as _jwt
 from auth_api.exceptions import BusinessException
-from auth_api.services import ResetTestData as ResetService
+from auth_api.services import Permissions as PermissionsService
 from auth_api.tracer import Tracer
 from auth_api.utils.endpoints_enums import EndpointEnum
-from auth_api.utils.roles import Role
 
-
-bp = Blueprint('RESET', __name__, url_prefix=f'{EndpointEnum.TEST_API.value}/reset')
+bp = Blueprint('MEMBER_PERMISSIONS', __name__, url_prefix=f'{EndpointEnum.API_V1.value}/permissions')
 TRACER = Tracer.get_instance()
 
 
-@bp.route('', methods=['POST'])
+@bp.route('/<string:org_status>/<string:membership_type>', methods=['GET', 'OPTIONS'])
 @TRACER.trace()
 @cross_origin(origin='*')
-@_jwt.has_one_of_roles([Role.TESTER.value])
-def post_reset():
-    """Cleanup test data by the provided token."""
+@_jwt.requires_auth
+def get_membership_permissions(org_status, membership_type):
+    """Get a list of all permissions for the membership."""
     try:
-        ResetService.reset()
-        response, status = '', http_status.HTTP_204_NO_CONTENT
+        case = request.args.get('case')
+        permissions = PermissionsService.get_permissions_for_membership(org_status.upper(), membership_type.upper())
+        if case == 'lower':
+            permissions = [x.lower() for x in permissions]
+        elif case == 'upper':
+            permissions = [x.upper() for x in permissions]
+
+        response, status = json.dumps(permissions), \
+            http_status.HTTP_200_OK
     except BusinessException as exception:
         response, status = {'code': exception.code, 'message': exception.message}, exception.status_code
     return response, status
